requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
crispy-bootstrap5==0.7
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.4
django-allauth==0.55.2
django-autoslug==1.9.9
django-browser-reload==1.11.0
django-ckeditor==6.7.0
django-crispy-forms==2.0
django-extensions==3.2.3
django-filter==23.2
django-formtools==2.4.1
django-js-asset==2.1.0
django-leaflet==0.29.0
django-phonenumber-field==7.1.0
django-taggit==4.0.0
flake8==6.1.0
idna==3.4
mccabe==0.7.0
oauthlib==3.2.2
phonenumberslite==8.13.19
Pillow==10.0.0
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
urllib3==2.0.4
